Mark fields as required automatically
We weren't showing clues about which fields are mandatory and which
fields are not. Sometimes we were showing the text "optional", but its
usage was inconsistent across our forms.

To mark fields as required, we've got a few options:

1. Add the HTML5 `required` attribute
2. Add a `<span title='Required'>*</span>` tag to the label
3. Use CSS to add content

The first two options are mostly incompatible with each other, since
screen reader users would hear the "required" text twice (once in the
label, once in the attribute). We would have to add an extra CSS rule
hiding the `<span>` tag to screen readers.

Another problem with the second option is we'd have to rewrite most our
specs dealing with forms, since the text in the labels has changed. So,
instead of `fill_in "Title"`, we would have to write `fill_in "Title*"`.

On the other hand, the first and third options complement each other.
Using CSS provides a visible cue for sighted users, while users of
screen readers will know the field is required thanks to the HTML5
attribute. The only disadvantage of this method is some sighted users
might not know what the `*` means (even if it is a de-facto standard in
web forms, not all users are necessarily aware of it). In any case, this
solution is much better than what we've done until now.

Note adding the `required` attribute will activate browser validations.
Since our forms haven't been designed with that validation in mind, it
would cause inconsistent behavior. So we're adding the `novalidate`
attribute to all our forms, in order to disable it. This is similar to
what the Simple Form used to do when browser validations are disabled
(they changed it nine years ago because back then browser support for
the `novalidate` attribute wasn't as good as it is today).

Also note adding the `required` attribute allows us to use the
`:required` CSS pseudoclass to style required inputs. However, since
labels usually appear before inputs (except in radio buttons and check
boxes), we can't use it to style labels and so we need to explicitely
add a `.required` class to the label.